updated delete and blacklist

+++ index.js
@@ -130,7 +130,6 @@ client.on('message', function (message) {
                     .channel
                     .send('Please register using `!ihl user register <username>` before trying to use this ' +
                             'bot');
-                message.delete();
             } else {
                 if (matches[1] && matches[1] === 'user') {
                     if (matches[2] && matches[2] === 'register') {
@@ -145,13 +144,11 @@ client.on('message', function (message) {
                         message
                             .channel
                             .send("Please specify a valid 'user' command");
-                        message.delete();
                     }
                 } else if (matches[1] && matches[1] === 'match') {
-                    // if(matches[2] && matches[2] === 'create') {
-                    // commands.match.create(message); } else if(matches[2] && matches[2] ===
-                    // 'join') {     commands.match.join(message, matches);     message.delete(); }
-                    // else
+                    // if(matches[2] && matches[2] === 'create') { commands.match.create(message); }
+                    // else if(matches[2] && matches[2] === 'join') { commands.match.join(message,
+                    // matches);     message.delete(); } else
                     if (matches[2] && matches[2] === 'report') {
                         commands
                             .match
@@ -160,14 +157,12 @@ client.on('message', function (message) {
                         message
                             .channel
                             .send("Please specify a valid 'match' command");
-                        message.delete();
                     }
                 } else if (matches[1] && matches[1] === 'queue') {
                     if (matches[2] && matches[2] === 'join') {
                         commands
                             .queue
                             .join(message, matches);
-                        message.delete();
                     } else if (matches[2] && matches[2] === 'leave') {
                         commands
                             .queue
@@ -179,10 +174,9 @@ client.on('message', function (message) {
                         .send("Please specify a valid command");
                 }
             }
-        } else {
-            if (message.author.id != '398933581314916362' && message.author.id != '121630407782432769') {
-                message.delete();
-            }
+        }
+        if (message.author.id != '398933581314916362' && message.author.id != '121630407782432769' && message.author.id != "207582905684197377") {
+            message.delete();
         }
     }
 });
